backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
opencv-python
onnxruntime
numpy
python-multipart
passlib[bcrypt]
Pillow
psutil
insightface
